requirements.txt
pillow==10.4.0
PyGithub==2.3.0
requests==2.32.3

